refactor(site-builder): deduplicate args pub & update
Deduplicate the arguments of Publish and Update
commands.

There was a duplication of some fields that were
shared among the publish and update.

The overlapping options were consolidated by
creating a PublishOptions struct that is then
added to both commands.
This new struct is also used in the publish_site
and update_site functions resolving the
"too many arguments" issue.