feat(portal): 139 create server side portal (#140)
Create a nextjs server that acts as a server-side walrus-sites portal.
It works by combining a middleware for preprocessing the requests, and a
root endpoint (defined in`route.ts`) that actually contains the logic as
the service worker did.

The middleware has the following role:
1. To handle requests like `http://<name>.walrus.site/path/to/file/`
without throwing an error. Explanation: if you don't define a route on
nextjs, the default logic is to throw an error because the
`/path/to/file` endpoint does not exist.
2. To delegate the whole request (including the subnames) to the
route.ts in order to fetch the site resources. By default the subnames
are removed from the hostname in nextjs.

In summary, the middleware is just a delegator and an error-preventer of
the original url.