Fix decompressData plaintext casting to Uint8Array
To fix it I had to create a new Uint8Array
instead of casting with `data as Uint8Array`.

> Analyzer error:
Conversion of type 'ArrayBuffer' to type 'Uint8Array'
 may be a mistake because neither type sufficiently
overlaps with the other. If this was intentional,
convert the expression to 'unknown' first.

Type 'ArrayBuffer' is missing the following
properties from type 'Uint8Array':
BYTES_PER_ELEMENT, buffer, byteOffset,
 copyWithin, and 29 more.