fix(portal): fix landing page server portal infinite loop (#148)
Resolves https://github.com/MystenLabs/walrus-sites/issues/147

Add the functionality so that while browsing the landing page of the
server portal fetch the page contents by the worker portal.

This is a temporary solution in order to make the server side portal
work.

As an improvement to this, in a following PR we can consider redirecting
to fetch the page from another service on vercel (similar to how the
worker portal works).