fix(portal): extract portal link (#203)
Closes #190 

Page 404 would provide a link for
walrus.site, even when the server side
portal is used (blob.store).

## Changes
Created a new function that extracts the 
appropriate redirection
link based on the current URL in order to
point to the correct portal, and uses it
to the `<a href=...>` tag.

i.e. if the user typed `<name>.walrus.site`
he should be redirected to `walrus.site` when
clicking the hyperlink.
Likewise, if the user typed `<name>.blob.store`, 
he should be redirected to `blob.store`.