Support any HTTP header
- Update the smart contract in order
to contain a VecMap of key (header name)
and value (header value) pairs.
Content-Type and Content-Encoding are no
longer defined explicitly in the contract.
They should be contained in the headers vec_map
as the rest of the headers.

- Update the site-builder so that
the right calls are made for to the
contract using the Rust SDK (thanks @giac-mysten)!

- Introduce a `ws-resources.json` file is used
to determine the HTTP response headers for each
file. This file location can also be specified
using the CLI args.