feat(contract): Add range field to Resource (#234)
This is used when multiple resources
share the same blob. To extract each
resource, the byte range field specifies
the slice of the blob to retrieve.

The aim of this change is to reach a stable 
smart contract version upon walrus testnet release. 

The corresponding site-builder and portal 
changes will be included in a separate PR.

---------

Signed-off-by: giac-mysten <giacomo@mystenlabs.com>
Co-authored-by: giac-mysten <giacomo@mystenlabs.com>