feat(common-lib): derive dynamic field instead of call to fullnode  (#246)
- Upgrade @mysten/sui to 1.12.0
This version includes the `deriveDynamicFieldID`
function needed to calculate the DF object ID
without making a request to a fullnode.
- Replace `client.getDynamicFieldObject` with `deriveDynamicFieldID`.
- Skip `page_fetching` benchmarks. The results we get from it are
misleading and slow down development whenever there are changes
happening in its' dependencies. In the near future we should revisit how
to properly bench this.
https://github.com/MystenLabs/walrus-sites/issues/250

TODO (following PR)
- ⚡ Performance improvement: Use `multiGetObjects` to fetch both the
display and the dynamic field object in the same request.