feat(portal): update portal domain name resolution using length (#293)
- Add a new `portalNameLength` wherever it applies.
- Update the `domain_parsing.tests` to cover the new functionality.

The PR is pointing to the `testnet` branch as it is the best branch to
test this.
It uses a `sw-tnet.blocksite.net` domain + the environment variables of
the portal name length have already been set.

Note: In the future, if we make the common library more object oriented,
this awkward argument passing can be avoided.