Create logger class
This class aims to be runtime-agnostic,
meaning that it should be able to run
in both the browser on on a server.

It is useful for cases where someone
needs to use a logging mechanism that
might be only supported on one runtime

For example amplitude provides different
interfaces (sdks) for different runtimes.