Create SENTRY_DSN env variable
This will be saved as a vercel
environment variable.

DSNs are safe to keep public because
they only allow submission of new
events and related event data;
they do not allow read access to any
information.

While there is a risk of abusing a DSN,
where any user can send events to your
organization with any information they want,
this is a rare occurrence.

So just to be safe 100%, we will use a vercel
environment variable for it.