feat(common): replace suix_resolveNameServiceAddress with SuinsClient.getNameRecord (#362)
### Summary

This PR integrates the new functionality of SuiNS name records
supporting a `walrusSiteId` to enable sites to connect with both a
wallet and a walrus site. For backwards compatibility, we fall back to
`nameRecord.targetAddress`.

Note: We don't need to use a `rpcSelector.validateSuinsResponse` because
if the full node fails to return something valid, it throws an error
that we catch on an outer try/catch block.

### Changelog
- install `@mysten/suins`
- upgrade `@mysten/sui` to `1.18.1` to be compatible with suins.
- add `getNameRecord` call to `WrappedSuiClient`.
- create a new `NameRecord` type in `common/types` since the suins lib
does not export it.
- update `SuiNSResolver` to use `rpcSelector.getNameRecord`.
- update unit tests accordingly.