feat: redis health check (#410)
When running `keyExists`, the code should not try to connect to redis
multiple when the connection is not ready.

This will cause unexpected redis connection connect & disconnect when a
pod is starting up, and multiple requests are coming in when the redis
connection is not fully ready - that explains a lot of corrupted pods
we've seen in kubernetes.

At the same time, I'm adding a `/api/healthz` endpoint for kubernetes
health check, so that it won't serve any traffic unless the pod is fully
ready.