feat(contract)!: sew 58 display object for walrus sites nft (#419)
## Summary

Devs can define a few `metadata` fields in `ws-resources.json` to be
included in the Site object's display.
This will make the Site objects better looking in explorers and easier
to spot inside a wallet.

The metadata are added during the site publishment. Currently if the
ws-resources.json get updated, the metadata will not be updated, but the
function `update_metadata` already exists in the contract. So a future
implementation can support this too.

Related walrus docs PR: 
~~https://github.com/MystenLabs/walrus-docs/pull/288~~
https://github.com/MystenLabs/walrus/pull/1758

> ⚠️ A new contract has to be published for this code to work!

> ❗ Only merge this PR when we decide to redeploy the contract! 

### Changelog

- Contract: Add more attributes to the `Site` object using a `Metadata`
struct.
- Contract: Define an `init` function and an `init_site_display`.
- Contract: Define a `Display` object for the `Site` objects. 
- Site builder: Update the site-builder's `create-site` function.
- Site builder: Parse the site metadata from `ws-resources.json`.
- Update the `examples/snake/ws-resources.json`.

### Extras
- Emit events upon site creation / deletion. 
- Fix the move-tests workflow: The actions would fail because of the
following:
```
2025-03-04T11:35:44.145272Z ERROR move_vm_runtime::logging: [VM] Unexpected verifier/deserialization error! This likely means there is code stored on chain that is unverifiable!
Error: VMError { major_status: MISSING_DEPENDENCY, sub_status: None, message: None, exec_state: None, location: Module(ModuleId { address: 0000000000000000000000000000000000000000000000000000000000000001, name: Identifier("unit_test") }), indices: [(FunctionHandle, 0)], offsets: [] }
[ FAIL    ] walrus_site::site_tests::test_new_range_both_bounds_defined
```

The issue here was in the removal of create_signers_for_testing native
function in this
[PR](https://github.com/MystenLabs/sui/pull/21372/files). That
framework/testnet still has the old framework which expects this native,
but it has been removed in the VM, so when it goes to load the old
packages it will fail on the new CLI.

The solution was to downgrade the Sui CLI to anything before v1.44
temporarily.

## Test

1. Publish a new contract.
2. Update the `sites-config.yaml` with the new packageId.
3. Publish the `examples/snake`.
4. Browse the site on suiscan. The Display object should take effect.

---------

Co-authored-by: giac-mysten <124184614+giac-mysten@users.noreply.github.com>