chore(portal): sew 234 add versions to sw and sp portals (#500)
#### Service Worker
The service worker and the server portal should each have a version
associated with it.
We don't host a service worker, and we don't put much effort into it to
be regarded as production ready, so I set it to 0.1.0.

#### Server Portal
The server portal on the other hand is production ready and battle
tested, so I set it to 1.0.0.