test(site-builder): Setup integration tests for e2e testing. (#495)
Introduced a lib.rs in site-builder to expose a Rust API for integration
testing.
- Moved Args from main.rs to the args.rs module.
- Retained an App wrapper in main.rs to support the bin_version!()
macro.

Added a tests/ directory with integration tests.
- localnode module defines TestSetup::start_local_test_cluster() to spin
up a Sui test cluster with Walrus and Walrus-Sites installed. It also
creates the required config files for Walrus and Sites, and new Wallet
and SuiClient objects for optional use in tests.
- convert, publish, and test_cluster test binaries perform basic tests
and serve as examples for writing future integration tests.