fix(site-builder): Update the site name if changed in ws-resources.json (#529)
When we updated the site name in the `ws-resources.json`, and then ran
`site-builder update ...` without any files being modified, then no
changes would occur. This is incorrect, the site name should be updated
as it is done with the site metadata.

Bonus: refactor the code so that the `SiteIdentifier` is replaced 
by `Option<ObjectID>`.