feat: use quilt patch ids in site-builder publish/update output (#561)
So far when we published a site, we would only show the blob_id to the
site builder output.

This PR covers the cases where we use the publish-quilts command. In
those cases we should
show the generated quilt patch id for each resource and not the blob id.

Bonus:
- Update `deny.toml` since some dependencies are no longer conflicting
with each other.