test(site-builder): Add test for quilts-lifecycle during site-builder updates (#581)
Add tests so that we make sure quilt-based site resources are always
extended despite the differences between the local and on chain version.

In detail, with the quilts functionality, Quilts are currently stored
before even comparing the local-resources with the site-resources. This
means that quilt-blobs are automatically extended without the need for
the `--check-extend` flag use.