chore(gh-actions): [SEW-380] Updates the migrated deploy github action (#601)
The deploy Github action has been migrated to a new repo for enhancing
security:
https://github.com/MystenLabs/walrus-sites-github-actions

The action in this repo redirects to the new action with a deprecation
note.
The example project used for showcasing this action has been also
migrated to the new repo.

In the context of this PR both changes here and the new repo should be
reviewed.