fix(site-builder): destroy sites with more than 1000 resources (#615)
Sites with over 1000 files were not being deleted due to this error:

```
Error: Exceeded maximum number of move-calls (1000) in Transaction
```

To solve this, I process the deleteResource move calls as resource
manager delete operations.
Then, the destroy_routes and burn site are two operations that are
appended last in the manager.
Operations are executed all together at last.